Moved the slices to the store/feature folder.
