+++ Moda.Web/src/moda.web.reactclient/src/app/components/common/health-check/create-health-check-form.tsx
@@ -12,7 +12,7 @@ import {
   selectHealthCheckEditContext,
   cancelHealthCheckCreate,
   getHealthCheckStatusOptions,
-} from '@/src/store/health-check-slice'
+} from '@/src/store/features/health-check-slice'
 import { useEffect } from 'react'
 
 const { Item } = Form
